+++ pages/setting.tsx
@@ -1,7 +1,7 @@
 import {Box} from '@chakra-ui/react';
 import type {NextPage} from 'next';
-import SettingForm from '../components/FormSetting/SettingForm';
 import Require from '../components/Session/Require';
+import SettingForm from '../components/Setting/Setting';
 
 const Setting: NextPage = () => {
   return (
